add uncrusrify-strong.cfg for unsatisfying but needed rules
Theses rules can make some people unhappy,
but there is not better alternative and doing
something is better than nothing.

for example it splits:

	VectorSubtract(vert[0]->normal, buildPlane->normal, normdiff); if (VectorLength(normdiff) >= normalEpsilon) {
		continue;
	}

as:

	VectorSubtract(vert[0]->normal, buildPlane->normal, normdiff);

	if (VectorLength(normdiff) >= normalEpsilon) {
		continue;
	}

There is no need for this extra blank line,
but we need the `if` in a new line, and there
is no option to add new line without blank line.